Migrate complex file transfer specs to os-integration-test (get)
