+++ server-integration-test/src/test/groovy/org/hidetake/groovy/ssh/test/server/AbstractFileTransferSpec.groovy
@@ -12,8 +12,6 @@ import spock.lang.Unroll
 import spock.util.concurrent.PollingConditions
 import spock.util.mop.Use
 
-import static org.hidetake.groovy.ssh.test.server.FileDivCategory.DirectoryType.DIRECTORIES
-import static org.hidetake.groovy.ssh.test.server.FileDivCategory.DirectoryType.DIRECTORY
 import static org.hidetake.groovy.ssh.test.server.FilenameUtils.toUnixPath
 
 @Use(FileDivCategory)
@@ -402,247 +400,6 @@ abstract class AbstractFileTransferSpec extends Specification {
         then:
         IOException e = thrown()
         e.message.contains(toUnixPath(sourceFile.path))
-    }
-
-
-
-    //
-    // [GET] file transfer
-    //
-
-    def "get(file) should create a file if destination is a non-existent file in a directory"() {
-        given:
-        def sourceFile = temporaryFolder.newFile() << 'Source Content'
-        def destinationFile = temporaryFolder.newFolder() / 'file1'
-        assert !destinationFile.exists()
-
-        when:
-        ssh.run {
-            session(ssh.remotes.testServer) {
-                get from: toUnixPath(sourceFile.path), into: destinationFile
-            }
-        }
-
-        then:
-        destinationFile.text == 'Source Content'
-    }
-
-    def "get(file) should create a file if destination is an existent directory"() {
-        given:
-        def sourceFile = temporaryFolder.newFile() << 'Source Content'
-
-        def destinationDir = temporaryFolder.newFolder()
-        def destination1File = destinationDir / sourceFile.name << 'Destination Content 1'
-        def destination2File = destinationDir / 'file2'         << 'Destination Content 2'
-
-        when:
-        ssh.run {
-            session(ssh.remotes.testServer) {
-                get from: toUnixPath(sourceFile.path), into: destinationDir
-            }
-        }
-
-        then: 'destination should be overwrote'
-        destination1File.text == 'Source Content'
-
-        and: 'destination should be kept as-is'
-        destination2File.text == 'Destination Content 2'
-    }
-
-    def "get(file) should overwrite a file if destination is an existent file"() {
-        given:
-        def sourceFile = temporaryFolder.newFile() << 'Source Content'
-        def destinationFile = temporaryFolder.newFile() << 'Destination Content'
-
-        when:
-        ssh.run {
-            session(ssh.remotes.testServer) {
-                get from: toUnixPath(sourceFile.path), into: destinationFile.path
-            }
-        }
-
-        then:
-        sourceFile.text == 'Source Content'
-        destinationFile.text == 'Source Content'
-    }
-
-    def "get(file) should throw IOException if destination and its parent do not exist"() {
-        given:
-        def sourceFile = temporaryFolder.newFile() << 'Source Content'
-        def destinationFile = temporaryFolder.newFolder() / 'dir1' / 'file1'
-
-        when:
-        ssh.run {
-            session(ssh.remotes.testServer) {
-                get from: toUnixPath(sourceFile.path), into: destinationFile
-            }
-        }
-
-        then:
-        IOException e = thrown()
-        e.message.contains(toUnixPath(destinationFile.path))
-    }
-
-
-
-    //
-    // [GET] directory transfer
-    //
-
-    def "get(dir) should create a directory if destination is an existent directory"() {
-        given:
-        def sourceDir = temporaryFolder.newFolder()
-        sourceDir / 'file1' << 'Source Content 1'
-        sourceDir / 'dir2' / DIRECTORY
-        sourceDir / 'dir2' / 'file2' << 'Source Content 2'
-        sourceDir / 'dir2' / 'dir3' / DIRECTORY
-
-        def destinationDir = temporaryFolder.newFolder()
-
-        when:
-        ssh.run {
-            session(ssh.remotes.testServer) {
-                get from: toUnixPath(sourceDir.path), into: destinationDir
-            }
-        }
-
-        then:
-        (destinationDir / sourceDir.name / 'file1').text == 'Source Content 1'
-        (destinationDir / sourceDir.name / 'dir2' / 'file2').text == 'Source Content 2'
-        (destinationDir / sourceDir.name / 'dir2' / 'dir3').list() == []
-    }
-
-    def "get(dir) should create a directory even if source is empty"() {
-        given:
-        def sourceDir = temporaryFolder.newFolder()
-        def destinationDir = temporaryFolder.newFolder()
-
-        when:
-        ssh.run {
-            session(ssh.remotes.testServer) {
-                get from: toUnixPath(sourceDir.path), into: destinationDir
-            }
-        }
-
-        then:
-        (destinationDir / sourceDir.name).list() == []
-    }
-
-    def "get(dir) should overwrite a directory if destination already exists"() {
-        given:
-        def sourceDir = temporaryFolder.newFolder()
-        sourceDir / 'file1' << 'Source Content 1'
-        sourceDir / 'dir2' / DIRECTORY
-        sourceDir / 'dir2' / 'file2' << 'Source Content 2'
-        sourceDir / 'dir2' / 'dir3' / DIRECTORY
-
-        and:
-        def destinationDir = temporaryFolder.newFolder()
-        destinationDir / sourceDir.name / DIRECTORY
-        destinationDir / sourceDir.name / 'file1' << 'Destination Content 1'
-        destinationDir / sourceDir.name / 'dir2' / DIRECTORY
-        destinationDir / sourceDir.name / 'dir2' / 'file2' << 'Destination Content 2'
-        destinationDir / sourceDir.name / 'dir2' / 'dir3' / DIRECTORY
-
-        when:
-        ssh.run {
-            session(ssh.remotes.testServer) {
-                get from: toUnixPath(sourceDir.path), into: destinationDir
-            }
-        }
-
-        then:
-        (destinationDir / sourceDir.name / 'file1').text == 'Source Content 1'
-        (destinationDir / sourceDir.name / 'dir2' / 'file2').text == 'Source Content 2'
-        (destinationDir / sourceDir.name / 'dir2' / 'dir3').list() == []
-    }
-
-    def "get(dir) should throw IOException if destination does not exist"() {
-        given:
-        def sourceDir = temporaryFolder.newFolder()
-        def destinationDir = temporaryFolder.newFolder() / 'dir1'
-        assert !destinationDir.exists()
-
-        when:
-        ssh.run {
-            session(ssh.remotes.testServer) {
-                get from: toUnixPath(sourceDir.path), into: destinationDir
-            }
-        }
-
-        then:
-        IOException e = thrown()
-        e.message.contains(toUnixPath(destinationDir.path))
-    }
-
-    def "get(dir) should get a directory recursively"() {
-        given:
-        def sourceDir = temporaryFolder.newFolder()
-        sourceDir / 'dir2' / 'dir3' / 'dir4' / 'dir5' / 'dir6' / 'dir7' / 'dir8' / 'dir9' / DIRECTORIES
-
-        sourceDir / 'file1' << 'Source Content 1'
-        sourceDir / 'dir2' / 'file2' << 'Source Content 2'
-        sourceDir / 'dir2' / 'dir3' / 'file3' << 'Source Content 3'
-        sourceDir / 'dir2' / 'dir3' / 'dir4' / 'file4' << 'Source Content 4'
-        sourceDir / 'dir2' / 'dir3' / 'dir4' / 'dir5' / 'file5' << 'Source Content 5'
-        sourceDir / 'dir2' / 'dir3' / 'dir4' / 'dir5' / 'dir6' / 'file6' << 'Source Content 6'
-        sourceDir / 'dir2' / 'dir3' / 'dir4' / 'dir5' / 'dir6' / 'dir7' / 'file7' << 'Source Content 7'
-        sourceDir / 'dir2' / 'dir3' / 'dir4' / 'dir5' / 'dir6' / 'dir7' / 'dir8' / 'file8' << 'Source Content 8'
-        sourceDir / 'dir2' / 'dir3' / 'dir4' / 'dir5' / 'dir6' / 'dir7' / 'dir8' / 'dir9' / 'file9' << 'Source Content 9'
-
-        def destinationDir = temporaryFolder.newFolder()
-
-        when:
-        ssh.run {
-            session(ssh.remotes.testServer) {
-                get from: toUnixPath(sourceDir.path), into: destinationDir
-            }
-        }
-
-        then:
-        (destinationDir / sourceDir.name / 'file1').text == 'Source Content 1'
-        (destinationDir / sourceDir.name / 'dir2' / 'file2').text == 'Source Content 2'
-        (destinationDir / sourceDir.name / 'dir2' / 'dir3' / 'file3').text == 'Source Content 3'
-        (destinationDir / sourceDir.name / 'dir2' / 'dir3' / 'dir4' / 'file4').text == 'Source Content 4'
-        (destinationDir / sourceDir.name / 'dir2' / 'dir3' / 'dir4' / 'dir5' / 'file5').text == 'Source Content 5'
-        (destinationDir / sourceDir.name / 'dir2' / 'dir3' / 'dir4' / 'dir5' / 'dir6' / 'file6').text == 'Source Content 6'
-        (destinationDir / sourceDir.name / 'dir2' / 'dir3' / 'dir4' / 'dir5' / 'dir6' / 'dir7' / 'file7').text == 'Source Content 7'
-        (destinationDir / sourceDir.name / 'dir2' / 'dir3' / 'dir4' / 'dir5' / 'dir6' / 'dir7' / 'dir8' / 'file8').text == 'Source Content 8'
-        (destinationDir / sourceDir.name / 'dir2' / 'dir3' / 'dir4' / 'dir5' / 'dir6' / 'dir7' / 'dir8' / 'dir9' / 'file9').text == 'Source Content 9'
-    }
-
-    @Unroll
-    def "get(dir) should get filtered files with regex #regex"() {
-        given:
-        def sourceDir = temporaryFolder.newFolder()
-        sourceDir / 'file1' << 'Source Content 1'
-        sourceDir / 'dir2' / DIRECTORY
-        sourceDir / 'dir2' / 'file2' << 'Source Content 2'
-        sourceDir / 'dir2' / 'dir3' / DIRECTORY
-
-        def destinationDir = temporaryFolder.newFolder()
-
-        when:
-        ssh.run {
-            session(ssh.remotes.testServer) {
-                get from: toUnixPath(sourceDir.path), into: destinationDir, filter: { it.name =~ regex }
-            }
-        }
-
-        then:
-        (destinationDir / sourceDir.name).exists() == d1
-        (destinationDir / sourceDir.name / 'file1').exists() == f1
-        (destinationDir / sourceDir.name / 'dir2').exists() == d2
-        (destinationDir / sourceDir.name / 'dir2' / 'file2').exists() == f2
-
-        and: 'empty directory should not be get'
-        !(destinationDir / sourceDir.name / 'dir2' / 'dir3').exists()
-
-        where:
-        regex | d1    | f1    | d2    | f2
-        /0$/  | false | false | false | false
-        /1$/  | true  | true  | false | false
-        /2$/  | true  | false | true  | true
     }
 
 }
